Update station list component and app to receive search filter
The search filter runs through a series of logic checks to determine
if any station info set in state matches what the user types. If it
does it renders those stations and reassigns the div styles in order
to keep the site style consistent.

Even with a list of only 20-30 radio stations it can be difficult for
a user to find exactly what station they want to listen to. This search
filter will enable the expansion of the API to include more stations
and generally increase the useability of the site.